Rework fetching to fetch only what's needed
That is, we fetch issues in ascending order by the date they
were last modified (hoping that this will include comments). We store
the comments first so we will always re-fetch an issue if there is
a problem storing the issue.

We stop fetching all issues in one go and will simply rely on this program
getting periodically restarted, slowly catching up with GitHub.